requirements.txt
python-telegram-bot[job-queue]~=21.10
google-api-python-client~=2.160
google-auth-httplib2~=0.2
google-auth-oauthlib~=1.2
google-genai~=1.5
python-dotenv~=1.1
pytz~=2024.2
